azmeta/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
adal==1.2.3               # via azure-cli-core, azure-kusto-data, msrestazure
applicationinsights==0.11.9  # via azure-cli-telemetry
argcomplete==1.11.1       # via azure-cli-core, knack
attrs==19.3.0             # via jsonschema
azure-cli-core==2.5.1     # via azmeta (setup.py)
azure-cli-nspkg==3.0.4    # via azure-cli-core, azure-cli-telemetry
azure-cli-telemetry==1.0.4  # via azure-cli-core
azure-common==1.1.25      # via azure-loganalytics, azure-mgmt-advisor, azure-mgmt-compute, azure-mgmt-consumption, azure-mgmt-resource, azure-mgmt-resourcegraph
azure-core==1.5.0         # via azmeta (setup.py), azure-mgmt-core
azure-kusto-data==0.0.45  # via azmeta (setup.py)
azure-loganalytics==0.1.0  # via azmeta (setup.py)
azure-mgmt-advisor==4.0.0  # via azmeta (setup.py)
azure-mgmt-compute==12.0.0  # via azmeta (setup.py)
azure-mgmt-consumption==3.0.0  # via azmeta (setup.py)
azure-mgmt-core==1.0.0    # via azure-cli-core
azure-mgmt-resource==9.0.0  # via azure-cli-core
azure-mgmt-resourcegraph==2.0.0  # via azmeta (setup.py)
azure-nspkg==3.0.2        # via azure-cli-nspkg, azure-loganalytics
bcrypt==3.1.7             # via paramiko
bleach==3.1.5             # via nbconvert
certifi==2020.4.5.1       # via msrest, requests
cffi==1.14.0              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via requests
colorama==0.4.3           # via azure-cli-core, knack
cryptography==2.9.2       # via adal, paramiko, pyjwt, pyopenssl
decorator==4.4.2          # via traitlets
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
humanfriendly==8.2        # via azure-cli-core
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via argcomplete, jsonschema
ipython-genutils==0.2.0   # via nbformat, traitlets
isodate==0.6.0            # via msrest
jinja2==2.11.2            # via nbconvert
jmespath==0.10.0          # via azure-cli-core, knack
jsonschema==3.2.0         # via nbformat
jupyter-core==4.6.3       # via nbconvert, nbformat
knack==0.7.0rc4           # via azure-cli-core
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via nbconvert
msal-extensions==0.1.3    # via azure-cli-core
msal==1.0.0               # via azure-cli-core, msal-extensions
msrest==0.6.13            # via azure-cli-core, azure-loganalytics, azure-mgmt-advisor, azure-mgmt-compute, azure-mgmt-consumption, azure-mgmt-resource, azure-mgmt-resourcegraph, msrestazure
msrestazure==0.6.3        # via azure-cli-core, azure-kusto-data, azure-mgmt-advisor, azure-mgmt-compute, azure-mgmt-consumption, azure-mgmt-resource, azure-mgmt-resourcegraph
nbconvert==5.6.1          # via azmeta (setup.py)
nbformat==5.0.6           # via nbconvert
numpy==1.18.4             # via pandas
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.3           # via bleach
pandas==1.0.3             # via azmeta (setup.py)
pandocfilters==1.4.2      # via nbconvert
paramiko==2.7.1           # via azure-cli-core
pkginfo==1.5.0.1          # via azure-cli-core
portalocker==1.7.0        # via azure-cli-telemetry, msal-extensions
pycparser==2.20           # via cffi
pygments==2.6.1           # via knack, nbconvert
pyjwt[crypto]==1.7.1      # via adal, azure-cli-core, msal
pynacl==1.3.0             # via paramiko
pyopenssl==19.1.0         # via azure-cli-core
pyparsing==2.4.7          # via packaging
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via adal, azure-kusto-data, pandas
pytz==2020.1              # via pandas
pyyaml==5.3.1             # via knack
requests-oauthlib==1.3.0  # via msrest
requests==2.23.0          # via adal, azure-cli-core, azure-core, azure-kusto-data, msal, msrest, requests-oauthlib
six==1.14.0               # via azure-cli-core, azure-core, bcrypt, bleach, cryptography, isodate, jsonschema, knack, packaging, pynacl, pyopenssl, pyrsistent, python-dateutil, traitlets
tabulate==0.8.7           # via knack
testpath==0.4.4           # via nbconvert
traitlets==4.3.3          # via jupyter-core, nbconvert, nbformat
urllib3==1.25.9           # via requests
webencodings==0.5.1       # via bleach
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
